requirements.txt
Flask==3.0.0
flask-cors==4.0.0
Pillow==10.1.0
google-genai>=0.2.0
Werkzeug==3.0.1
requests==2.31.0
python-dotenv==1.0.0
Flask-Login==0.6.3
Flask-Bcrypt==1.0.1
Flask-SQLAlchemy==3.1.1
Authlib==1.3.0
stripe==7.8.0
gunicorn==21.2.0

